Fixed process id button bug.

+++ hello-world.py
@@ -7,30 +7,21 @@ from os import kill
 import subprocess
 import signal
 
-# ~ window = Gtk.Window(title="Hello World!")
-# ~ window.connect("destroy", Gtk.main_quit)
-# ~ window.show()
-# ~ Gtk.main()
-
-
-##Bug: When start button is pressed multiple times, the pid gets updated
-##so the stop button will have the wrong id to kill
-##The music continues to run under the first pid.
-
 class Handler:
 	myProcess = None
 	
 	def button1_clicked(self, button):
 		print("Start music button pressed!")
 		if self.myProcess is None:
-			self.myProcess = subprocess.Popen(["ogg123", "james-brown-dead.ogg", "-r"])
+			self.myProcess = subprocess.Popen(["ogg123", "james-brown-dead.ogg", "-r", "-q"])
 			print("Process ID: " + str(self.myProcess.pid))
 		
 	def music_stop_clicked_cb(self, button):
 		print("Stop button pressed!")
-		print("Killing this process: " + str(self.myProcess.pid))
-		kill(self.myProcess.pid, signal.SIGKILL)
-		self.myProcess = None
+		if self.myProcess is not None:
+			print("Killing this process: " + str(self.myProcess.pid))
+			kill(self.myProcess.pid, signal.SIGKILL)
+			self.myProcess = None
 
 builder = Gtk.Builder()
 builder.add_from_file("glade-button.glade")
